+++ lib/widgets/numeric_fields/card_settings_double.dart
@@ -37,7 +37,7 @@ class CardSettingsDouble extends CardSettingsText {
     bool maxLengthEnforced: true,
     ValueChanged<String> onFieldSubmitted,
     List<TextInputFormatter> inputFormatters,
-    bool showMaterialonIOS: false,
+    bool showMaterialonIOS,
   }) : super(
           key: key,
           label: label,