+++ lib/widgets/picker_fields/card_settings_list_picker.dart
@@ -28,7 +28,7 @@ class CardSettingsListPicker extends FormField<String> {
     this.hintText,
     this.options,
     this.values,
-    this.showMaterialonIOS = false,
+    this.showMaterialonIOS,
   })  : assert(values == null || options.length == values.length,
             "If you provide 'values', they need the same number as 'options'"),
         super(
@@ -81,7 +81,7 @@ class _CardSettingsListPickerState extends FormFieldState<String> {
     } else {
       optionIndex = 0; // set to first element in the list
     }
-    if (showCupertino(widget.showMaterialonIOS))
+    if (showCupertino(context, widget.showMaterialonIOS))
       _showCupertinoBottomPicker(optionIndex);
     else
       _showMaterialScrollPicker(label, option);
@@ -175,7 +175,7 @@ class _CardSettingsListPickerState extends FormFieldState<String> {
       content = options[optionIndex];
     }
 
-    if (showCupertino(widget.showMaterialonIOS))
+    if (showCupertino(context, widget.showMaterialonIOS))
       return cupertinoSettingsListPicker(content);
     else
       return materialSettingsListPicker(content);