+++ lib/widgets/picker_fields/card_settings_selection_picker.dart
@@ -29,7 +29,7 @@ class CardSettingsSelectionPicker extends FormField<String> {
     this.options,
     this.icons,
     this.values,
-    this.showMaterialonIOS = false,
+    this.showMaterialonIOS,
   })  : assert(values == null || options.length == values.length,
             "If you provide 'values', they need the same number as 'options'"),
         assert(values == null || options.length == icons.length,
@@ -87,7 +87,7 @@ class _CardSettingsListPickerState extends FormFieldState<String> {
     } else {
       optionIndex = 0; // set to first element in the list
     }
-    if (showCupertino(widget.showMaterialonIOS))
+    if (showCupertino(context, widget.showMaterialonIOS))
       _showCupertinoBottomPicker(optionIndex);
     else
       _showMaterialSelectionPicker(label, option);
@@ -182,7 +182,7 @@ class _CardSettingsListPickerState extends FormFieldState<String> {
       content = options[optionIndex];
     }
 
-    if (showCupertino(widget.showMaterialonIOS))
+    if (showCupertino(context, widget.showMaterialonIOS))
       return cupertinoSettingsListPicker(content);
     else
       return materialSettingsListPicker(content);