+++ lib/widgets/text_fields/card_settings_paragraph.dart
@@ -37,7 +37,7 @@ class CardSettingsParagraph extends CardSettingsText {
     bool maxLengthEnforced: true,
     ValueChanged<String> onFieldSubmitted,
     List<TextInputFormatter> inputFormatters,
-    bool showMaterialonIOS: false,
+    bool showMaterialonIOS,
   }) : super(
           key: key,
           label: label,